requirements.txt
pyspark 
psycopg2-binary
streamlit
pandas